8716 - K8s Proxy Charms
This commit enables the possibility to deploy
Proxy Charms on Kubernetes.

It depends on a new LCM environment variable
called `k8s_cloud` which will state the k8s name
in VCA that will be used for deploying the proxy
charms.

It will also check the k8s model for status
and delete the k8s model if it exists.

Finally it will now retry the verify ssh
credential action when the action fails, instead
of only retrying when a excpetion is launched.

Change-Id: I09510dcec00f747da945f88d54c2e25c1402d56c
Signed-off-by: Dominik Fleischmann <dominik.fleischmann@canonical.com>
(cherry picked from commit b95133489d1cd16e1692085b1193d24d52c858fc)