fix bugs 1117 and 1142 at deleting k8scluster helm-chart
bug 1142: When helm is not going to be uninstalled, do not check for kdu releases

bug 1117: When helm is going to be unistalled with kdu releases, just warn and
skip uninstall helm. This happens because kdu release has been installed
outside the scope of OSM. NBI already checks that there is not OSM kdu
deployments upon k8scluster deletion, denying the operation if any.

Change-Id: Icee2125b790c562130e13dd8c619d1c94c64c0fe
Signed-off-by: tierno <alfonso.tiernosepulveda@telefonica.com>